detection-as-code-phase3/pyproject.toml
[project]
name = "detection-as-code"
version = "0.3.0"
description = "Sigma-based detection engineering pipeline with CI validation, automated conversion, and continuous detection validation against a home SOC lab via Atomic Red Team detonation."
authors = [{ name = "Samim Hakimi" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
  "pysigma>=0.11.0",
  "pysigma-backend-splunk>=1.1.0",
  "pysigma-backend-elasticsearch>=1.1.0",
  "pyyaml>=6.0",
  "requests>=2.31",
  "pywinrm>=0.4.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/sammyHa/detection-as-code"
Issues = "https://github.com/sammyHa/detection-as-code/issues"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra"
